# contracts/verification-control-plane/openapi.yaml
openapi: 3.1.0
info:
  title: Verification Control Plane API
  version: 0.1.0
  description: Provider-neutral API for evidence intake, verification lifecycle, human review, adjudication, verdicts, and agent feedback.
servers:
  - url: https://verification-control-plane.example.test
    description: Placeholder staging endpoint for contract validation
paths:
  /verification-jobs:
    post:
      summary: Create or retrieve a verification job
      operationId: createVerificationJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVerificationJobRequest"
      responses:
        "202":
          description: Job accepted or existing active job returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationJob"
        "400":
          description: Invalid or non-observable acceptance criteria
        "409":
          description: Conflicting idempotency key with incompatible payload
  /verification-jobs/{jobId}:
    get:
      summary: Get verification job status
      operationId: getVerificationJob
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "200":
          description: Current job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationJob"
        "404":
          description: Job not found
...
  /verification-jobs/{jobId}/release-artifact:
    get:
      summary: Signed release artifact for a finalized job (operator)
      operationId: getReleaseArtifact
      tags: [operator]
      parameters:
        - $ref: "#/components/parameters/jobId"
        - $ref: "#/components/parameters/operatorToken"
      responses:
        "200":
          description: Signed release artifact
        "404":
          description: Artifact not available
components:
  parameters:
    jobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
...
    operatorToken:
      name: x-operator-token
      in: header
      required: false
      schema:
        type: string
      description: Required when RUNTIME_OPERATOR_TOKEN is configured
  schemas:
    CreateVerificationJobRequest:
      type: object
      required:
        [
          idempotency_key,
          source,
          risk_tier,
          acceptance_criteria,
          deadline_at,
          budget_policy
        ]
      properties:
        idempotency_key:
          type: string
        agent_run_id:
          type: string
        parent_job_id:
          type: string
        source:
          $ref: "#/components/schemas/JobSource"
        risk_tier:
          $ref: "#/components/schemas/RiskTier"
        acceptance_criteria:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AcceptanceCriterion"
        deadline_at:
          type: string
          format: date-time
        budget_policy:
          $ref: "#/components/schemas/BudgetPolicy"
    VerificationJob:
      allOf:
        - $ref: "#/components/schemas/CreateVerificationJobRequest"
        - type: object
          required: [job_id, state, created_at, updated_at]
          properties:
            job_id:
              type: string
            state:
              $ref: "#/components/schemas/JobState"
            artifact_manifest_id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    JobSource:
      type: object
      required: [repository, commit, environment, route]
      properties:
        repository:
          type: string
        branch:
          type: string
        commit:
          type: string
        build_id:
          type: string
        environment:
          type: string
        route:
          type: string
        viewport:
          type: string
        locale:
          type: string
    RiskTier:
      type: string
      enum: [low, medium, high, regulated, release_gating]
    JobState:
      type: string
      enum:
        - created
        - artifacts_collected
        - privacy_classified
        - self_verifying
        - decision_point
        - external_review_queued
        - internal_review_queued
        - human_responses_received
        - consensus_running
        - adjudication_required
        - final_pass
        - final_fail
        - agent_retry_requested
        - artifact_recapture_requested
        - fail_closed
        - canceled
    AcceptanceCriterion:
      type: object
      required:
        [criterion_id, human_visible_text, criticality, evidence_requirements]
      properties:
        criterion_id:
          type: string
        human_visible_text:
          type: string
        criticality:
          type: string
          enum: [critical, major, minor, audit]
        evidence_requirements:
          type: array
          items:
            type: string
        pass_threshold:
          type: number
          minimum: 0
          maximum: 1
    BudgetPolicy:
      type: object
      required: [max_job_cost, max_assignments, max_retries]
      properties:
        max_job_cost:
          type: number
          minimum: 0
        max_assignments:
          type: integer
          minimum: 0
        max_retries:
          type: integer
          minimum: 0
...